Clean up lint, typing, security, and migrations

+++ src/kaleta/services/data_service.py
@@ -3,6 +3,7 @@
 Used by the Settings page for the "Populate with example data" and
 "Clear all data" developer/demo actions.
 """
+
 from __future__ import annotations
 
 import datetime
@@ -32,35 +33,78 @@ _YEARS = 6
 _MONTHS = _YEARS * 12
 
 _EXPENSE_CATEGORIES = [
-    "Żywność", "Restauracje & Kawiarnie", "Transport", "Paliwo",
-    "Mieszkanie & Czynsz", "Media (prąd, gaz, woda)", "Zdrowie & Apteka",
-    "Rozrywka", "Odzież & Obuwie", "Elektronika", "Sport & Fitness",
-    "Edukacja", "Subskrypcje", "Wakacje & Podróże", "Inne wydatki",
+    "Żywność",
+    "Restauracje & Kawiarnie",
+    "Transport",
+    "Paliwo",
+    "Mieszkanie & Czynsz",
+    "Media (prąd, gaz, woda)",
+    "Zdrowie & Apteka",
+    "Rozrywka",
+    "Odzież & Obuwie",
+    "Elektronika",
+    "Sport & Fitness",
+    "Edukacja",
+    "Subskrypcje",
+    "Wakacje & Podróże",
+    "Inne wydatki",
 ]
 _INCOME_CATEGORIES = ["Wynagrodzenie", "Freelance", "Zwroty", "Inne przychody"]
 _BUDGETED_CATEGORIES = [
-    "Żywność", "Mieszkanie & Czynsz", "Transport", "Media (prąd, gaz, woda)",
-    "Zdrowie & Apteka", "Rozrywka", "Subskrypcje", "Odzież & Obuwie",
+    "Żywność",
+    "Mieszkanie & Czynsz",
+    "Transport",
+    "Media (prąd, gaz, woda)",
+    "Zdrowie & Apteka",
+    "Rozrywka",
+    "Subskrypcje",
+    "Odzież & Obuwie",
 ]
 _BASE_BUDGETS: dict[str, Decimal] = {
-    "Żywność":                  Decimal("1400.00"),
-    "Mieszkanie & Czynsz":      Decimal("2400.00"),
-    "Transport":                Decimal("350.00"),
-    "Media (prąd, gaz, woda)":  Decimal("280.00"),
-    "Zdrowie & Apteka":         Decimal("200.00"),
-    "Rozrywka":                 Decimal("300.00"),
-    "Subskrypcje":              Decimal("120.00"),
-    "Odzież & Obuwie":          Decimal("250.00"),
+    "Żywność": Decimal("1400.00"),
+    "Mieszkanie & Czynsz": Decimal("2400.00"),
+    "Transport": Decimal("350.00"),
+    "Media (prąd, gaz, woda)": Decimal("280.00"),
+    "Zdrowie & Apteka": Decimal("200.00"),
+    "Rozrywka": Decimal("300.00"),
+    "Subskrypcje": Decimal("120.00"),
+    "Odzież & Obuwie": Decimal("250.00"),
 }
 _SEASONAL: dict[int, float] = {
-    1: 0.75, 2: 0.85, 3: 0.90, 4: 0.95, 5: 1.00, 6: 1.05,
-    7: 1.25, 8: 1.30, 9: 1.05, 10: 0.95, 11: 1.10, 12: 1.50,
+    1: 0.75,
+    2: 0.85,
+    3: 0.90,
+    4: 0.95,
+    5: 1.00,
+    6: 1.05,
+    7: 1.25,
+    8: 1.30,
+    9: 1.05,
+    10: 0.95,
+    11: 1.10,
+    12: 1.50,
 }
 _CATCH_PHRASES = [
-    "Zakupy spożywcze", "Paliwo", "Kino / Netflix", "Apteka", "Restauracja",
-    "Odzież", "Kosmetyki", "Elektronika", "Siłownia", "Książki",
-    "Kawiarnia", "Bilety", "Parking", "Fryzjer", "Leki",
-    "Zabawki", "Ogród", "Materiały biurowe", "Prezent", "Usługi",
+    "Zakupy spożywcze",
+    "Paliwo",
+    "Kino / Netflix",
+    "Apteka",
+    "Restauracja",
+    "Odzież",
+    "Kosmetyki",
+    "Elektronika",
+    "Siłownia",
+    "Książki",
+    "Kawiarnia",
+    "Bilety",
+    "Parking",
+    "Fryzjer",
+    "Leki",
+    "Zabawki",
+    "Ogród",
+    "Materiały biurowe",
+    "Prezent",
+    "Usługi",
 ]
 
 
@@ -70,12 +114,12 @@ def _month_offset(today: datetime.date, n: int) -> tuple[int, int]:
 
 
 def _inflation(months_back: int, annual: float = 0.045) -> float:
-    return 1.0 / ((1 + annual) ** (months_back / 12))
+    return float(1.0 / ((1 + annual) ** (months_back / 12)))
 
 
 def _salary(months_back: int) -> Decimal:
     base = 9000.0 * _inflation(months_back, annual=0.05)
-    return Decimal(str(round(base * random.uniform(0.95, 1.05), 2)))
+    return Decimal(str(round(base * random.uniform(0.95, 1.05), 2)))  # nosec B311
 
 
 # ── service ───────────────────────────────────────────────────────────────────
@@ -113,25 +157,31 @@ class DataService:
 
     async def seed(self) -> dict[str, int]:
         """Clear all data then insert 6 years of realistic Polish-language demo data."""
-        rng = random.Random(42)
+        rng = random.Random(42)  # nosec B311: deterministic demo-data generator, not security.
         await self.clear_all()
         s = self.session
 
         # ── institutions ──────────────────────────────────────────────────────
         institutions = [
             Institution(
-                name="PKO Bank Polski", type=InstitutionType.BANK,
-                color="#003087", website="https://www.pkobp.pl",
+                name="PKO Bank Polski",
+                type=InstitutionType.BANK,
+                color="#003087",
+                website="https://www.pkobp.pl",
                 description="Największy bank w Polsce",
             ),
             Institution(
-                name="mBank", type=InstitutionType.FINTECH,
-                color="#e2001a", website="https://www.mbank.pl",
+                name="mBank",
+                type=InstitutionType.FINTECH,
+                color="#e2001a",
+                website="https://www.mbank.pl",
                 description="Nowoczesny bank internetowy",
             ),
             Institution(
-                name="Revolut", type=InstitutionType.FINTECH,
-                color="#191c1f", website="https://www.revolut.com",
+                name="Revolut",
+                type=InstitutionType.FINTECH,
+                color="#191c1f",
+                website="https://www.revolut.com",
                 description="Aplikacja finansowa — karty i wymiana walut",
             ),
         ]
@@ -141,14 +191,25 @@ class DataService:
 
         # ── accounts ──────────────────────────────────────────────────────────
         accounts = [
-            Account(name="PKO Konto Główne",     type=AccountType.CHECKING,
-                    balance=Decimal("0.00"), institution_id=pko.id),
-            Account(name="mBank Oszczędności",   type=AccountType.SAVINGS,
-                    balance=Decimal("0.00"), institution_id=mbank.id),
-            Account(name="Gotówka",              type=AccountType.CASH,
-                    balance=Decimal("0.00")),
-            Account(name="Karta Kredytowa Visa", type=AccountType.CREDIT,
-                    balance=Decimal("0.00"), institution_id=revolut.id),
+            Account(
+                name="PKO Konto Główne",
+                type=AccountType.CHECKING,
+                balance=Decimal("0.00"),
+                institution_id=pko.id,
+            ),
+            Account(
+                name="mBank Oszczędności",
+                type=AccountType.SAVINGS,
+                balance=Decimal("0.00"),
+                institution_id=mbank.id,
+            ),
+            Account(name="Gotówka", type=AccountType.CASH, balance=Decimal("0.00")),
+            Account(
+                name="Karta Kredytowa Visa",
+                type=AccountType.CREDIT,
+                balance=Decimal("0.00"),
+                institution_id=revolut.id,
+            ),
         ]
         s.add_all(accounts)
         await s.flush()
@@ -156,15 +217,15 @@ class DataService:
 
         # ── categories ────────────────────────────────────────────────────────
         expense_cats = [Category(name=n, type=CategoryType.EXPENSE) for n in _EXPENSE_CATEGORIES]
-        income_cats  = [Category(name=n, type=CategoryType.INCOME)  for n in _INCOME_CATEGORIES]
+        income_cats = [Category(name=n, type=CategoryType.INCOME) for n in _INCOME_CATEGORIES]
         s.add_all(expense_cats + income_cats)
         await s.flush()
 
-        by_name      = {c.name: c for c in expense_cats + income_cats}
-        salary_cat   = by_name["Wynagrodzenie"]
+        by_name = {c.name: c for c in expense_cats + income_cats}
+        salary_cat = by_name["Wynagrodzenie"]
         freelance_cat = by_name["Freelance"]
-        zwroty_cat   = by_name["Zwroty"]
-        rent_cat     = by_name["Mieszkanie & Czynsz"]
+        zwroty_cat = by_name["Zwroty"]
+        rent_cat = by_name["Mieszkanie & Czynsz"]
         vacation_cat = by_name["Wakacje & Podróże"]
         electronics_cat = by_name["Elektronika"]
 
@@ -186,84 +247,109 @@ class DataService:
             seasonal = _SEASONAL[month]
             inf = _inflation(m)
 
-            add_tx(Transaction(
-                account_id=checking.id, category_id=salary_cat.id,
-                amount=_salary(m), type=TransactionType.INCOME,
-                date=datetime.date(year, month, 1),
-                description=f"Wynagrodzenie {month:02d}/{year}",
-            ))
+            add_tx(
+                Transaction(
+                    account_id=checking.id,
+                    category_id=salary_cat.id,
+                    amount=_salary(m),
+                    type=TransactionType.INCOME,
+                    date=datetime.date(year, month, 1),
+                    description=f"Wynagrodzenie {month:02d}/{year}",
+                )
+            )
 
-            rent_amount = Decimal(str(round(
-                float(_BASE_BUDGETS["Mieszkanie & Czynsz"]) * inf, 2
-            )))
-            add_tx(Transaction(
-                account_id=checking.id, category_id=rent_cat.id,
-                amount=rent_amount, type=TransactionType.EXPENSE,
-                date=datetime.date(year, month, 5),
-                description="Czynsz za mieszkanie",
-            ))
+            rent_amount = Decimal(str(round(float(_BASE_BUDGETS["Mieszkanie & Czynsz"]) * inf, 2)))
+            add_tx(
+                Transaction(
+                    account_id=checking.id,
+                    category_id=rent_cat.id,
+                    amount=rent_amount,
+                    type=TransactionType.EXPENSE,
+                    date=datetime.date(year, month, 5),
+                    description="Czynsz za mieszkanie",
+                )
+            )
 
             for _ in range(rng.randint(12, 22)):
                 cat = rng.choice(expense_cats)
-                amount = Decimal(str(round(
-                    rng.uniform(8, 600) * seasonal * inf, 2
-                )))
-                add_tx(Transaction(
-                    account_id=rng.choice([checking, cash, credit]).id,
-                    category_id=cat.id,
-                    amount=amount, type=TransactionType.EXPENSE,
-                    date=datetime.date(year, month, rng.randint(1, 28)),
-                    description=rng.choice(_CATCH_PHRASES),
-                ))
+                amount = Decimal(str(round(rng.uniform(8, 600) * seasonal * inf, 2)))
+                add_tx(
+                    Transaction(
+                        account_id=rng.choice([checking, cash, credit]).id,
+                        category_id=cat.id,
+                        amount=amount,
+                        type=TransactionType.EXPENSE,
+                        date=datetime.date(year, month, rng.randint(1, 28)),
+                        description=rng.choice(_CATCH_PHRASES),
+                    )
+                )
 
             if month in (7, 8) and rng.random() < 0.6:
-                add_tx(Transaction(
-                    account_id=checking.id, category_id=vacation_cat.id,
-                    amount=Decimal(str(round(rng.uniform(1500, 5000) * inf, 2))),
-                    type=TransactionType.EXPENSE,
-                    date=datetime.date(year, month, rng.randint(1, 20)),
-                    description="Wyjazd wakacyjny",
-                ))
+                add_tx(
+                    Transaction(
+                        account_id=checking.id,
+                        category_id=vacation_cat.id,
+                        amount=Decimal(str(round(rng.uniform(1500, 5000) * inf, 2))),
+                        type=TransactionType.EXPENSE,
+                        date=datetime.date(year, month, rng.randint(1, 20)),
+                        description="Wyjazd wakacyjny",
+                    )
+                )
 
             if month == 12 and rng.random() < 0.5:
-                add_tx(Transaction(
-                    account_id=credit.id, category_id=electronics_cat.id,
-                    amount=Decimal(str(round(rng.uniform(800, 3500) * inf, 2))),
-                    type=TransactionType.EXPENSE,
-                    date=datetime.date(year, month, rng.randint(10, 23)),
-                    description="Prezenty świąteczne / elektronika",
-                ))
+                add_tx(
+                    Transaction(
+                        account_id=credit.id,
+                        category_id=electronics_cat.id,
+                        amount=Decimal(str(round(rng.uniform(800, 3500) * inf, 2))),
+                        type=TransactionType.EXPENSE,
+                        date=datetime.date(year, month, rng.randint(10, 23)),
+                        description="Prezenty świąteczne / elektronika",
+                    )
+                )
 
             if rng.random() < 0.35:
-                add_tx(Transaction(
-                    account_id=checking.id, category_id=freelance_cat.id,
-                    amount=Decimal(str(round(rng.uniform(400, 4000) * inf, 2))),
-                    type=TransactionType.INCOME,
-                    date=datetime.date(year, month, rng.randint(10, 25)),
-                    description="Faktura freelance",
-                ))
+                add_tx(
+                    Transaction(
+                        account_id=checking.id,
+                        category_id=freelance_cat.id,
+                        amount=Decimal(str(round(rng.uniform(400, 4000) * inf, 2))),
+                        type=TransactionType.INCOME,
+                        date=datetime.date(year, month, rng.randint(10, 25)),
+                        description="Faktura freelance",
+                    )
+                )
 
             if rng.random() < 0.15:
-                add_tx(Transaction(
-                    account_id=checking.id, category_id=zwroty_cat.id,
-                    amount=Decimal(str(round(rng.uniform(20, 300) * inf, 2))),
-                    type=TransactionType.INCOME,
-                    date=datetime.date(year, month, rng.randint(1, 28)),
-                    description="Zwrot / reklamacja",
-                ))
+                add_tx(
+                    Transaction(
+                        account_id=checking.id,
+                        category_id=zwroty_cat.id,
+                        amount=Decimal(str(round(rng.uniform(20, 300) * inf, 2))),
+                        type=TransactionType.INCOME,
+                        date=datetime.date(year, month, rng.randint(1, 28)),
+                        description="Zwrot / reklamacja",
+                    )
+                )
 
             # internal transfer checking → savings
             t_amount = Decimal(str(round(rng.uniform(300, 1500) * inf, 2)))
             t_date = datetime.date(year, month, 15)
             t_out = Transaction(
-                account_id=checking.id, category_id=None,
-                amount=t_amount, type=TransactionType.TRANSFER, date=t_date,
+                account_id=checking.id,
+                category_id=None,
+                amount=t_amount,
+                type=TransactionType.TRANSFER,
+                date=t_date,
                 description=f"Przelew własny → oszczędności {month:02d}/{year}",
                 is_internal_transfer=True,
             )
             t_in = Transaction(
-                account_id=savings.id, category_id=None,
-                amount=t_amount, type=TransactionType.TRANSFER, date=t_date,
+                account_id=savings.id,
+                category_id=None,
+                amount=t_amount,
+                type=TransactionType.TRANSFER,
+                date=t_date,
                 description=f"Przelew własny ← konto główne {month:02d}/{year}",
                 is_internal_transfer=True,
             )
@@ -276,11 +362,14 @@ class DataService:
             balance_delta[savings.id] += t_amount
 
             for cat_name in _BUDGETED_CATEGORIES:
-                all_budgets.append(Budget(
-                    category_id=by_name[cat_name].id,
-                    amount=Decimal(str(round(float(_BASE_BUDGETS[cat_name]) * inf, 2))),
-                    month=month, year=year,
-                ))
+                all_budgets.append(
+                    Budget(
+                        category_id=by_name[cat_name].id,
+                        amount=Decimal(str(round(float(_BASE_BUDGETS[cat_name]) * inf, 2))),
+                        month=month,
+                        year=year,
+                    )
+                )
 
         s.add_all(all_tx)
         s.add_all(all_budgets)
@@ -291,21 +380,24 @@ class DataService:
         # ── physical assets ───────────────────────────────────────────────────
         assets = [
             Asset(
-                name="Mieszkanie (Warszawa)", type=AssetType.REAL_ESTATE,
+                name="Mieszkanie (Warszawa)",
+                type=AssetType.REAL_ESTATE,
                 value=Decimal("620000.00"),
                 description="Mieszkanie 52m² na Mokotowie, zakupione w 2021",
                 purchase_date=datetime.date(2021, 6, 15),
                 purchase_price=Decimal("480000.00"),
             ),
             Asset(
-                name="Toyota Corolla 2020", type=AssetType.VEHICLE,
+                name="Toyota Corolla 2020",
+                type=AssetType.VEHICLE,
                 value=Decimal("68000.00"),
                 description="Toyota Corolla Hybrid 1.8, rok 2020",
                 purchase_date=datetime.date(2020, 3, 10),
                 purchase_price=Decimal("95000.00"),
             ),
             Asset(
-                name="Zegarek Seiko", type=AssetType.VALUABLES,
+                name="Zegarek Seiko",
+                type=AssetType.VALUABLES,
                 value=Decimal("4500.00"),
                 description="Seiko Prospex, edycja limitowana",
             ),
